+++ Fabric/src/main/java/vazkii/botania/fabric/client/FabricPlatformModel.java
@@ -19,7 +19,6 @@ import net.minecraft.util.RandomSource;
 import net.minecraft.world.level.BlockAndTintGetter;
 import net.minecraft.world.level.block.state.BlockState;
 
-import vazkii.botania.common.block.BotaniaBlocks;
 import vazkii.botania.common.block.PlatformBlock;
 import vazkii.botania.common.block.block_entity.PlatformBlockEntity;
 
@@ -52,11 +51,6 @@ public class FabricPlatformModel extends ForwardingBakedModel {
 				// No camo
 				super.emitBlockQuads(blockView, state, pos, randomSupplier, context);
 			} else {
-				// Some people used this to get an invisible block in the past, accommodate that.
-				if (heldState.is(BotaniaBlocks.manaGlass)) {
-					return;
-				}
-
 				BakedModel model = Minecraft.getInstance().getBlockRenderer()
 						.getBlockModelShaper().getBlockModel(heldState);
 				if (model instanceof FabricBakedModel) {